requirements/doc.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
accessible-pygments==0.0.5
    # via pydata-sphinx-theme
alabaster==1.0.0
    # via sphinx
asgiref==3.8.1
    # via
    #   -r requirements/test.txt
    #   django
babel==2.16.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
backports-tarfile==1.2.0
    # via jaraco-context
beautifulsoup4==4.12.3
    # via pydata-sphinx-theme
build==1.2.1
    # via -r requirements/doc.in
cachetools==5.5.0
    # via
    #   -r requirements/test.txt
    #   tox
certifi==2024.7.4
    # via
    #   -r requirements/test.txt
    #   elasticsearch
    #   requests
chardet==5.2.0
    # via
    #   -r requirements/test.txt
    #   tox
charset-normalizer==3.3.2
    # via
    #   -r requirements/test.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/test.txt
    #   code-annotations
code-annotations==1.8.0
    # via -r requirements/test.txt
colorama==0.4.6
    # via
    #   -r requirements/test.txt
    #   tox
coverage[toml]==7.6.1
    # via
    #   -r requirements/test.txt
    #   pytest-cov
distlib==0.3.8
    # via
    #   -r requirements/test.txt
    #   virtualenv
django==4.2.15
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements/test.txt
dnspython==2.6.1
    # via
    #   -r requirements/test.txt
    #   pymongo
doc8==1.1.1
    # via -r requirements/doc.in
docutils==0.20.1
    # via
    #   doc8
    #   pydata-sphinx-theme
    #   readme-renderer
    #   restructuredtext-lint
    #   sphinx
elasticsearch==7.13.4
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.txt
filelock==3.15.4
    # via
    #   -r requirements/test.txt
    #   tox
    #   virtualenv
idna==3.7
    # via
    #   -r requirements/test.txt
    #   requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.2.0
    # via
    #   keyring
    #   twine
iniconfig==2.0.0
    # via
    #   -r requirements/test.txt
    #   pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==5.3.0
    # via keyring
jaraco-functools==4.0.2
    # via keyring
jinja2==3.1.4
    # via
    #   -r requirements/test.txt
    #   code-annotations
    #   sphinx
keyring==25.3.0
    # via twine
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   -r requirements/test.txt
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
mongomock==4.1.2
    # via -r requirements/test.txt
more-itertools==10.4.0
    # via
    #   jaraco-classes
    #   jaraco-functools
nh3==0.2.18
    # via readme-renderer
openedx-atlas==0.6.1
    # via -r requirements/test.txt
packaging==24.1
    # via
    #   -r requirements/test.txt
    #   build
    #   mongomock
    #   pydata-sphinx-theme
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
pbr==6.0.0
    # via
    #   -r requirements/test.txt
    #   stevedore
pkginfo==1.10.0
    # via twine
platformdirs==4.2.2
    # via
    #   -r requirements/test.txt
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   -r requirements/test.txt
    #   pytest
    #   tox
pydata-sphinx-theme==0.15.4
    # via sphinx-book-theme
pygments==2.18.0
    # via
    #   accessible-pygments
    #   doc8
    #   pydata-sphinx-theme
    #   readme-renderer
    #   rich
    #   sphinx
pymongo==4.8.0
    # via -r requirements/test.txt
pyproject-api==1.7.1
    # via
    #   -r requirements/test.txt
    #   tox
pyproject-hooks==1.1.0
    # via build
pytest==8.3.2
    # via
    #   -r requirements/test.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.txt
pytest-django==4.8.0
    # via -r requirements/test.txt
python-slugify==8.0.4
    # via
    #   -r requirements/test.txt
    #   code-annotations
pyyaml==6.0.2
    # via
    #   -r requirements/test.txt
    #   code-annotations
readme-renderer==43.0
    # via twine
requests==2.32.3
    # via
    #   -r requirements/test.txt
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==1.0.0
    # via twine
restructuredtext-lint==1.4.0
    # via doc8
rfc3986==2.0.0
    # via twine
rich==13.7.1
    # via twine
sentinels==1.0.0
    # via
    #   -r requirements/test.txt
    #   mongomock
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.6
    # via beautifulsoup4
sphinx==8.0.2
    # via
    #   -r requirements/doc.in
    #   pydata-sphinx-theme
    #   sphinx-book-theme
sphinx-book-theme==1.1.3
    # via -r requirements/doc.in
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlparse==0.5.1
    # via
    #   -r requirements/test.txt
    #   django
stevedore==5.2.0
    # via
    #   -r requirements/test.txt
    #   code-annotations
    #   doc8
text-unidecode==1.3
    # via
    #   -r requirements/test.txt
    #   python-slugify
tox==4.18.0
    # via -r requirements/test.txt
twine==5.1.1
    # via -r requirements/doc.in
typing-extensions==4.12.2
    # via pydata-sphinx-theme
urllib3==1.26.19
    # via
    #   -r requirements/test.txt
    #   elasticsearch
    #   requests
    #   twine
virtualenv==20.26.3
    # via
    #   -r requirements/test.txt
    #   tox
zipp==3.20.0
    # via importlib-metadata
